Docker_Inventory_Management-main/consumer_3/requirements.txt
pika==1.3.2
mysql-connector-python==8.2.0
flask==3.0.2

